редактирование ДЗ
