+++ index.html
@@ -236,6 +236,6 @@
     </template>
 
 </body>
-<script src="./js/main.js" type="module"></script>
+<script src="js/main.js" type="module"></script>
 <script src="./js/functions.js"></script>
 </html>
